+++ lib/src/data/core/api_service.dart
@@ -2,7 +2,8 @@ import 'package:dio/dio.dart';
 import 'package:talker_dio_logger/talker_dio_logger_interceptor.dart';
 import 'package:talker_dio_logger/talker_dio_logger_settings.dart';
 import 'package:youth_biomarkers_sdk/src/data/core/rest_client.dart';
-import 'package:youth_biomarkers_sdk/src/data/models/haut_analyze_model.dart';
+import 'package:youth_biomarkers_sdk/src/data/models/haut_analyze_marker_model.dart';
+import 'package:youth_biomarkers_sdk/src/data/models/haut_analyze_payload_data.dart';
 
 class ApiService {
   static final ApiService _instance = ApiService._();
@@ -20,7 +21,9 @@ class ApiService {
     return _instance;
   }
 
-  Future<HautAnalyzeModel> getSelfieAnalyzeData(String imageBase64) async {
-    return _client.analyzeSelfieByHaut(imageBase64);
+  Future<List<HautAnalyzeMarkerModel>> getSelfieAnalyzeData(
+      String imageBase64) async {
+    return _client.analyzeSelfieByHaut(
+        "1", "678961acc7fc9ac83a0e26f0", HautAnalyzePayloadData(imageBase64));
   }
 }