+++ lib/src/data/models/haut_analyze_payload_data.g.dart
@@ -1,0 +1,19 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'haut_analyze_payload_data.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+HautAnalyzePayloadData _$HautAnalyzePayloadDataFromJson(
+        Map<String, dynamic> json) =>
+    HautAnalyzePayloadData(
+      json['base64url_bytes'] as String,
+    );
+
+Map<String, dynamic> _$HautAnalyzePayloadDataToJson(
+        HautAnalyzePayloadData instance) =>
+    <String, dynamic>{
+      'base64url_bytes': instance.base64Image,
+    };